files_manage/requirements.txt
django
djangorestframework
mysqlclient